dumb bug in the file auth store impl

+++ crates/jacquard-common/src/session.rs
@@ -94,7 +94,9 @@ impl FileTokenStore {
     /// Create a new file token store at the given path.
     pub fn new(path: impl AsRef<Path>) -> Self {
         std::fs::create_dir_all(path.as_ref().parent().unwrap()).unwrap();
-        std::fs::write(path.as_ref(), b"{}").unwrap();
+        if !path.as_ref().exists() {
+            std::fs::write(path.as_ref(), b"{}").unwrap();
+        }
 
         Self {
             path: path.as_ref().to_path_buf(),
